% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbfar.R
\name{nbfar_control}
\alias{nbfar_control}
\title{Default control parameters for Generalized co-sparse factor regresion}
\usage{
nbfar_control(
  maxit = 5000,
  epsilon = 1e-06,
  elnetAlpha = 0.95,
  gamma0 = 1,
  spU = 0.5,
  spV = 0.5,
  lamMaxFac = 1,
  lamMinFac = 1e-06,
  initmaxit = 2000,
  initepsilon = 1e-06,
  objI = 1
)
}
\arguments{
\item{maxit}{maximum iteration for each sequential steps}

\item{epsilon}{tolerence value set for convergene of gcure}

\item{elnetAlpha}{elastic net penalty parameter}

\item{gamma0}{power parameter in the adaptive weights}

\item{spU}{maximum proportion of nonzero elements in each column of U}

\item{spV}{maximum proportion of nonzero elements in each column of V}

\item{lamMaxFac}{a multiplier of calculated lambda_max}

\item{lamMinFac}{a multiplier of determing lambda_min as a fraction of lambda_max}

\item{initmaxit}{maximum iteration for initialization problem}

\item{initepsilon}{tolerence value for convergene in the initialization problem}

\item{objI}{1 or 0 convergence on the basis of objective function or not}
}
\value{
a list of controling parameter.
}
\description{
Control parameters for NBFAR
}
\examples{
control <- nbfar_control()
}
