% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulate_nbfar}
\alias{simulate_nbfar}
\title{Simulated data for NBFAR}
\format{
A dlist of variables for the analysis  using NBFAR and NBRRR:
\describe{
\item{Y}{Generated response matrix}
\item{X}{Generated predictor matrix}
\item{U}{specified value of U}
\item{V}{specified value of V}
\item{D}{specified value of D}
\item{n}{sample size}
\item{Xsigma}{covariance matrix used to generate predictors in X}
\item{C0}{intercept value in the coefficient matrix}
\item{disp}{dispersion parameter of the generative model}
\item{depth}{log of the sequencing depth of the microbiome data (used as an offset in the simulated multivariate negative binomial regression model)}
Mishra, A., Müller, C. (2022) Negative binomial factor regression models with application to microbiome data analysis.  https://doi.org/10.1101/2021.11.29.470304}
}
\usage{
data(simulate_nbfar)
}
\description{
Simulated data with low-rank and sparse coefficient matrix.
}
\keyword{datasets}
