% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbfar.R
\name{nbrrr}
\alias{nbrrr}
\title{Negative binomial reduced rank regression (NBRRR)}
\usage{
nbrrr(
  Yt,
  X,
  maxrank = 10,
  cIndex = NULL,
  ofset = "CSS",
  control = list(),
  nfold = 5,
  trace = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{Yt}{response matrix}

\item{X}{design matrix; when X = NULL, we set X as identity matrix and perform generalized PCA.}

\item{maxrank}{an integer specifying the maximum possible rank of the coefficient matrix or the number of factors}

\item{cIndex}{specify index of control variable in the design matrix X}

\item{ofset}{offset matrix or microbiome data analysis specific scaling: common sum scaling = CSS (default), total sum scaling = TSS, median-ratio scaling = MRS, centered-log-ratio scaling  = CLR}

\item{control}{a list of internal parameters controlling the model fitting}

\item{nfold}{number of folds in k-fold crossvalidation}

\item{trace}{TRUE/FALSE checking progress of cross validation error}

\item{verbose}{TRUE/FALSE checking progress of estimation procedure}
}
\value{
\item{C}{estimated coefficient matrix}
\item{Z}{estimated control variable coefficient matrix}
\item{PHI}{estimted dispersion parameters}
\item{U}{estimated U matrix (generalize latent factor weights)}
\item{D}{estimated singular values}
\item{V}{estimated V matrix (factor loadings)}
}
\description{
In the range of 1 to maxrank, the estimation procedure selects the rank r of the coefficient matrix using a cross-validation approach. For the selected rank, a rank r coefficient matrix is estimated that best fits the observations.
}
\examples{
\donttest{
## Load simulated data set:
data('simulate_nbfar')
attach(simulate_nbfar)

# Model with known offset
set.seed(1234)
offset <- log(10)*matrix(1,n,ncol(Y))
control_nbrr <- nbfar_control(initmaxit = 5000, initepsilon = 1e-4)
# nbrrr_test <- nbrrr(Y, X, maxrank = 5, cIndex = NULL, ofset = offset,
#                       control = control_nbrr, nfold = 5)
}
}
\references{
Mishra, A., Müller, C. (2022) \emph{Negative binomial factor regression models with application to microbiome data analysis.  https://doi.org/10.1101/2021.11.29.470304}
}
